requirements.txt
flask
gunicorn
google-generativeai
werkzeug
python-docx